requirements.txt
allure-pytest==2.13.5
allure_python_commons==2.13.5
names==0.3.0
pillow==11.2.1
pyhumps==3.8.0
pytesseract==0.3.13
pytest==8.3.4
selenium==4.32.0
pytest-xdist==3.6.1
pytest-dependency==0.6.0
pytest-env==1.1.5
tensorflow==2.20.0
numpy==2.3.0
scikit-learn==1.7.2
opencv-python==4.11.0.86
